requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-modeltranslation==0.19.12
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
